BudanTables

Documentation for BudanTables.

BudanTables.budantable

BudanTables.set_getting_roots

BudanTables.budantable — Function

Plots the Budan's table of a given polynomial (see https://hal.inria.fr/hal-00653762/document). The input is a polynomial of type T for which 1.- degree(p::T) and derivative(p::T, i::Int) must be defined 2.- it is compatible with function computing roots f (see help of BudanTables.set_getting_roots).

By default it uses T==Polynomial from Polynomials.jl.

Examples of use:

Using Polynomials' types and api:

   using BudanTables
   using Plots

   using Polynomials

   pp = Polynomial([-1,2,3])
   budantable(pp)

Or define our own api for julia vectors:

   using BudanTables
   using Plots

   BudanTables.degree(v::AbstractVector) = length(v) - 1

   BudanTables.derivative(v::AbstractVector) = .*(v[begin+1:end], 1:(length(v)-1))
   function BudanTables.derivative(v::AbstractVector, i::Int)
       out = copy(v)
       while i > 0
           i-=1
           out = derivative(out)
       end
       return out
   end

   # PolynomialRoots.jl represents polynomials as julia vectors.
   import PolynomialRoots as PR
   BudanTables.set_getting_roots(PR.roots)

   pp = [-1,-2,3,4]
   budantable(pp)

BudanTables.set_getting_roots — Method

set_getting_roots(f)

Sets the function used to compute the roots of polynomials building a Budan's table. The input is a function f computing the roots of a polynomial represented as an object of type T, for which degree(p::T) and derivative(p::T) must be defined. By default it uses roots from Polynomials.jl.

An example of valid input for polynomials of type Polynomial is the following:

   using Polynomials
   pp = Polynomial([-1,2,3])
   import PolynomialRoots as PR # Fast roots
   # Notice pre-composition with Polynomials' `coeffs`.
   BudanTables.set_getting_roots(PR.roots∘coeffs)
   budantable(pp)