BudanTables

Documentation for BudanTables.

BudanTables.budantable

BudanTables.set_getting_roots

BudanTables.budantable — Function

Plots the Budan's table of a polynomial (see https://hal.inria.fr/hal-00653762/document). The input is a polynomial of type T for which

degree(p::T) and derivative(p::T, i::Int) must be defined

it is compatible with function computing roots (see set_getting_roots).

The default are type T==Polynomial and Polynomials.roots from Polynomials.jl. But, if PolynomialRoots.jl is loaded, it uses PolynomialRoots.roots (accepting polynomials of type Polynomial).

Examples

Default use, type Polynomials and roots Polynomials.roots

   using Plots
   using Polynomials
   using BudanTables
   P = fromroots([0,0,1,2,2,4,4])
   budantable(P)

Use type Polynomials and roots PolynomialRoots.roots

   using Plots
   using Polynomials
   using PolynomialRoots
   using BudanTables
   P = fromroots([0,0,1,2,2,4,4])
   budantable(P)

Or define our own API for julia vectors:

   using Plots
   using BudanTables

   # Setting function to compute roots from a polynomial represented
   # as a vector.
   import PolynomialRoots
   _BigFloat(A::AbstractArray) = BigFloat.(A)
   # Notice we do not precompose by `coeffs`
   set_getting_roots(PolynomialRoots.roots∘_BigFloat)

   BudanTables.degree(v::AbstractVector) = length(v) - 1
   BudanTables.derivative(v::AbstractVector) = .*(v[begin+1:end], 1:(length(v)-1))
   function BudanTables.derivative(v::AbstractVector, i::Int)
       out = copy(v)
       while i > 0
           i-=1
           out = derivative(out)
       end
       return out
   end

   import Polynomials
   P = Polynomials.coeffs(Polynomials.fromroots([0,0,1,2,2,4,4]))
   budantable(P)

BudanTables.set_getting_roots — Method

set_getting_roots(f)

Sets the function used to compute the roots of polynomials building a Budan's table. The input is a function f computing the roots of a polynomial represented as an object of type T, for which degree(p::T) and derivative(p::T) must be defined.

See also budantable